requirements.txt
-i https://pypi.python.org/simple
asgiref==3.6.0 ; python_version >= '3.7'
click==8.1.3 ; python_version >= '3.7'
django==4.1.5
django-environ==0.9.0
h11==0.14.0 ; python_version >= '3.7'
psycopg2==2.9.5
sqlparse==0.4.3 ; python_version >= '3.5'
uvicorn==0.20.0
